+++ changelog.md
@@ -3,3 +3,4 @@
   - Links any 2 entities together 
 - Add chipped mushrooms to bouncy tag
 - Fix Dango Blossom missing from Building Blocks Creative Tab
+- Fix Plastic Buckets unable to bucket water spouts